display list of upcoming and past launches
