--- src/app/components/launches-list/launches-list.component.html
@@ -1,3 +1,0 @@
-<p>
-  launches-list works!
-</p>
