--- src/app/components/launches-list/launches-list.component.ts
@@ -1,15 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'app-launches-list',
-  templateUrl: './launches-list.component.html',
-  styleUrls: ['./launches-list.component.css']
-})
-export class LaunchesListComponent implements OnInit {
-
-  constructor() { }
-
-  ngOnInit() {
-  }
-
-}
