+++ src/app/modules/app-routing/app-routing.module.ts
@@ -2,9 +2,11 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { RouterModule, Routes } from '@angular/router';
 import { HomeComponent } from '../../pages/home/home.component';
+import { LaunchesListComponent } from '../../pages/launches-list/launches-list.component';
 
 const appRoutes: Routes = [
   { path: '', component: HomeComponent },
+  { path: 'launches', component: LaunchesListComponent },
 ];
 
 @NgModule({
