+++ src/app/pages/launches-list/launches-list.component.html
@@ -1,0 +1,11 @@
+<ul>
+  <li *ngFor="let launch of launchesUpcomning; let i = index">
+    {{launch.mission_name}} {{launch.mission_name}}
+  </li>
+</ul>
+
+<ul>
+  <li *ngFor="let launch of launchesPast; let i = index">
+    {{launch.mission_name}} {{launch.mission_name}}
+  </li>
+</ul>
